Add load previous history to helper functions
by creating fetch request like done in UIKit, the function is able to sort by a variable given to it and then it only returns the first element of the fetch request. Which should be the closest history. Allowing the sets to be grabbed from previous week and compared to current week.